Inject browser scripts when needed
Inject browser scripts in the `MockService.setup()` function, but only
if they are needed. This way, it is easy to install the `MockService`
and we don't have to rely on `onPageLoad()` from `lib/index.js` (which
isn't called in most environments).

It is now possible to use protractor-xmlhttprequest-mock like this:

    import {browser, $} from 'protractor';
    import {MockService} from 'protractor-xmlhttprequest-mock';

    describe('the backend', () => {
      it('should reply with code 418', () => {
        MockService.setup(browser);
        MockService.addMock('mock1', {
          path: '/api/teapot',
          response: {status: 418, data: "I'm a teapot"},
        });

        $('button').click();
        expect($('div').getText()).toEqual("I'm a teapot");
      });
    });